+++ src/models/submit_order_market_dto_data.rs
@@ -32,7 +32,7 @@ pub struct SubmitOrderMarketDtoData {
     pub side: models::OrderSide,
     /// Onchain generated productId from prior product registration
     #[serde(rename = "onchainId")]
-    pub onchain_id: f64,
+    pub onchain_id: i32,
     #[serde(rename = "engineType")]
     pub engine_type: models::EngineType,
     /// A subaccount scoped unique client-generated order id (either a UUID or alphanumeric string up to 32 characters)
@@ -52,10 +52,10 @@ pub struct SubmitOrderMarketDtoData {
     pub stop_type: Option<StopType>,
     /// Message signedAt current timestamp (seconds since Unix Epoch)
     #[serde(rename = "signedAt")]
-    pub signed_at: f64,
+    pub signed_at: i64,
     /// Order expiry timestamp (seconds since Unix Epoch), defaults to the maximum allowed value: signedAt + 6652800
     #[serde(rename = "expiresAt", skip_serializing_if = "Option::is_none")]
-    pub expires_at: Option<f64>,
+    pub expires_at: Option<i64>,
     /// Group Id (UUID) for linking orders together in OCO/OTO relationships
     #[serde(rename = "groupId", skip_serializing_if = "Option::is_none")]
     pub group_id: Option<uuid::Uuid>,
@@ -75,9 +75,9 @@ impl SubmitOrderMarketDtoData {
         r#type: Type,
         quantity: String,
         side: models::OrderSide,
-        onchain_id: f64,
+        onchain_id: i32,
         engine_type: models::EngineType,
-        signed_at: f64,
+        signed_at: i64,
     ) -> SubmitOrderMarketDtoData {
         SubmitOrderMarketDtoData {
             subaccount,
